move debug commands to archiver.debug
